# Improve performance by switching to partial search