This updates the chef node searches to use partial searches which should help serialization and memory usage locally.

```bash
# Before
bundle exec knife inventory chef_client  16.52s user 0.93s system 82% cpu 21.159 total

#After
bundle exec knife inventory chef_client  0.69s user 0.16s system 8% cpu 10.479 total
```